--- R_workspace/ams-raw_fitted-values.csv
@@ -1,38 +1,0 @@
-,0,1
-0,30.339498817920685,28.702076121886964
-1,31.553999662399292,31.386291298938296
-2,33.51850017905235,33.216557639297456
-3,33.921000599861145,34.67201745212602
-4,34.063998656347394,35.91387861488321
-5,34.435500210151076,37.01771606774233
-6,38.01899814605713,38.02561572134125
-7,38.063001573085785,38.96371626216165
-8,38.06850108504295,39.849554861408556
-9,39.230900049209595,40.69561293722072
-10,40.583999663591385,41.511210473046525
-11,42.38500029966235,42.303599473501464
-12,42.60109984874725,43.07863580296371
-13,43.927998542785645,43.841214634618915
-14,45.487000942230225,44.59556685947713
-15,45.94700127840042,45.3454709028803
-16,47.582002490758896,46.09441218545123
-17,47.787998735904694,46.84571046319196
-18,47.95249950885773,47.60262865354307
-19,47.97799950838089,48.36847316437115
-20,48.203497767448425,49.146694045639194
-21,49.76050353050232,49.94099295877686
-22,49.949998915195465,50.75544787396414
-23,52.79249954223633,51.59466574969299
-24,54.16449884325266,52.463978794378974
-25,54.233500957489014,53.36970743129159
-26,54.24599742889404,54.319526063439326
-27,56.30400061607361,55.322990642525276
-28,56.635998614132404,56.39232905169236
-29,58.23269961774349,57.543676242719144
-30,58.370999336242676,58.799101861371724
-31,61.75579750537872,60.19014455224003
-32,61.801498889923096,61.76445870464218
-33,62.78049898147583,63.59963814730552
-34,69.50500160455704,65.83633780382306
-35,69.53780148923397,68.77699456398369
-36,74.06800197064877,73.32611510237578
