--- R_workspace/ams_24H_duration.csv
@@ -1,38 +1,0 @@
-time,Anchorage,Barrow,Bethel,Fairbanks,Juneau,Nome
-1979-12-31,35.790000915527344,10.119999885559082,21.40949821472168,19.725000381469727,114.11100006103516,18.191999435424805
-1980-12-31,30.92900085449219,11.934999465942385,24.11999893188477,42.9119987487793,77.9800033569336,19.647998809814453
-1981-12-31,25.355998992919922,11.183500289916992,18.518999099731445,18.215999603271484,118.01799011230469,26.38399887084961
-1982-12-31,36.90999984741211,13.545999526977539,20.503999710083008,26.391998291015625,83.61599731445312,31.649999618530273
-1983-12-31,24.125999450683594,7.11299991607666,21.545000076293945,15.567499160766602,71.98699951171875,22.420000076293945
-1984-12-31,25.5,12.050000190734865,14.41800022125244,18.51300048828125,60.95299911499024,23.3439998626709
-1985-12-31,31.141000747680664,19.916000366210934,26.809999465942386,23.768001556396484,58.660499572753906,43.257999420166016
-1986-12-31,63.65999603271485,11.58400058746338,18.287500381469727,30.011999130249023,63.42399978637695,38.90299987792969
-1987-12-31,19.751998901367188,8.715999603271483,44.17649841308594,31.191999435424805,67.60800170898439,25.51799964904785
-1988-12-31,25.56149864196777,12.868000030517578,21.145999908447266,19.742000579833984,112.90750885009766,22.4060001373291
-1989-12-31,21.65900039672852,11.98799991607666,19.027999877929688,29.1460018157959,68.79199981689453,24.889501571655273
-1990-12-31,29.717998504638672,13.742000579833984,30.71750068664551,29.24250030517578,98.13249969482422,22.93400001525879
-1991-12-31,19.57349967956543,10.482000350952148,21.29599952697754,25.507999420166016,76.62600708007812,29.30200004577637
-1992-12-31,37.3380012512207,10.713000297546387,32.28900146484375,18.73350143432617,53.16999816894531,25.7760009765625
-1993-12-31,20.663000106811523,21.934999465942386,25.58799934387207,22.86199951171875,133.16000366210938,24.993999481201172
-1994-12-31,30.39799880981445,13.908499717712402,36.652000427246094,26.559999465942386,112.09999084472656,43.43000030517578
-1995-12-31,20.45800018310547,13.697500228881836,22.608999252319336,18.26099967956543,69.35950469970703,25.17599868774414
-1996-12-31,20.86000061035156,9.920000076293944,28.17700004577637,15.18299961090088,113.95800018310548,33.150001525878906
-1997-12-31,21.902000427246094,11.621500015258787,22.678998947143555,13.699000358581545,75.13999938964844,20.66750144958496
-1998-12-31,29.09000015258789,17.798500061035156,24.1560001373291,22.12900161743164,77.02999877929688,38.24799728393555
-1999-12-31,54.82049942016602,11.868999481201172,17.27400016784668,16.108999252319336,80.96499633789062,25.399999618530273
-2000-12-31,20.40049934387207,13.15999984741211,23.189998626708984,14.952500343322756,86.260498046875,19.94599914550781
-2001-12-31,17.61050033569336,18.80500030517578,20.23699951171875,14.968999862670898,77.84549713134766,21.215999603271484
-2002-12-31,14.376999855041506,19.46199989318848,16.49100112915039,27.197998046875,122.21400451660156,20.839500427246094
-2003-12-31,23.81599998474121,16.139999389648438,18.38599967956543,43.233497619628906,75.87200164794922,24.34250068664551
-2004-12-31,36.834999084472656,16.54800033569336,21.03849983215332,16.625999450683597,86.67549896240234,25.127500534057614
-2005-12-31,29.053001403808594,14.782500267028807,15.420000076293945,31.663000106811523,87.23999786376953,25.538000106811523
-2006-12-31,34.12200164794922,11.96399974822998,26.40199851989746,15.619999885559082,84.26300048828125,35.553001403808594
-2007-12-31,31.290000915527344,9.845000267028807,17.101999282836914,45.36600112915039,65.46649932861328,41.62799835205078
-2008-12-31,25.4325008392334,7.622499465942383,20.440000534057614,27.54999923706055,66.63700103759766,30.27299880981445
-2009-12-31,37.15999984741211,11.889999389648438,15.42549991607666,28.25,102.68999481201172,23.608501434326172
-2010-12-31,17.040000915527347,12.632999420166016,26.25199890136719,17.98200035095215,71.00399780273438,16.145999908447266
-2011-12-31,29.333999633789066,12.373000144958494,45.2750015258789,16.420000076293945,101.10499572753906,22.792499542236328
-2012-12-31,35.060001373291016,8.359999656677246,30.48200035095215,18.269001007080078,57.00100326538086,26.62399864196777
-2013-12-31,37.95549774169922,9.910999298095703,23.45000076293945,18.850500106811523,67.39399719238281,22.860998153686523
-2014-12-31,26.145999908447266,15.420000076293945,14.59600067138672,29.479999542236328,82.13749694824219,19.753000259399414
-2015-12-31,21.11400032043457,16.600000381469727,20.31800079345703,28.53000259399414,94.8499984741211,29.41250038146973
