--- R_workspace/ams_24H_duration_fbx.csv
@@ -1,37 +1,0 @@
-1979,0.7765752375185012
-1980,1.6894496819393157
-1981,0.7171657259807587
-1982,1.0390556119171142
-1983,0.6128939987092972
-1984,0.7288586805236816
-1985,0.9357485980754852
-1986,1.181575406957817
-1987,1.2280321369726181
-1988,0.7772445370281219
-1989,1.147481006088066
-1990,1.151280161264801
-1991,1.004252487971878
-1992,0.7375398248195646
-1993,0.9000792069763183
-1994,1.0456698349740983
-1995,0.7189373834844589
-1996,0.5977562129811287
-1997,0.5393310140173913
-1998,0.8712210065784454
-1999,0.6342129114637375
-2000,0.5886814337666512
-2001,0.5893310214933395
-2002,1.0707879029052734
-2003,1.7021071246345518
-2004,0.6545672609733583
-2005,1.2465754805051803
-2006,0.6149609574944496
-2007,1.7860640010547637
-2008,1.0846462249629976
-2009,1.112205325
-2010,0.7079531520170211
-2011,0.6464570450037003
-2012,0.7192523965488433
-2013,0.7421460742551803
-2014,1.1606305299777984
-2015,1.1232290551258086
