--- R_workspace/fbx_7day_ams_era-interim.csv
@@ -1,38 +1,0 @@
-time,Fairbanks
-1979,1.319626703899309
-1980,2.292627207220523
-1981,1.677209561155164
-1982,1.897776527454221
-1983,1.8888986384449065
-1984,1.8878947009136795
-1985,1.4985441782325446
-1986,2.132461735908741
-1987,2.229764929038254
-1988,2.2980720809678075
-1989,2.7377101984112904
-1990,1.8733081962614266
-1991,2.4716745229506016
-1992,1.5445244580273867
-1993,1.729449695429325
-1994,1.6687017002977367
-1995,1.597796125155559
-1996,1.2422841221086263
-1997,1.8089380350307522
-1998,1.4987606945682495
-1999,1.3557290868236689
-2000,2.1356703433753013
-2001,1.3411030335002625
-2002,2.431331923366511
-2003,2.433131191446161
-2004,1.1944691024114191
-2005,2.2166941346548796
-2006,1.8814182890324413
-2007,2.916064644384639
-2008,2.0784459862277984
-2009,1.9665364522911368
-2010,1.4968117589100838
-2011,1.790827775795698
-2012,2.135178356046438
-2013,1.3354731857165931
-2014,2.7364188636715707
-2015,1.9590760000462293
