--- ancillary/alaska_test_locations.cpg
@@ -1,1 +1,0 @@
-ISO-8859-1
