+++ readme.md
@@ -6,7 +6,7 @@ These data constitute an effort to update the NOAA Atlas 14 for Alaska using mor
 
 The data transformation in this project follows a pipeline with the following steps: 
 
-1. **Starting point:** WRF hourly pcpt data (can be obtained [here](http://wrf-ak-ar5.s3-website-us-east-1.amazonaws.com/)), along with the NOAA Atlas 14 data.
+1. **Starting point:** WRF hourly pcpt data (can be obtained [here](http://wrf-ak-ar5.s3-website-us-east-1.amazonaws.com/)), along with the NOAA Atlas 14 data. (See the scripts in [here](other_scripts/) for help obtaining and pre-processing the NOAA Atlas 14 data).
 2. **Durations:** Calculate duration series for various time periods.
 3. **AMS:** Calculate **A**nnual **M**aximum **S**eries for all duration series.
 4. **Intervals:** Calculate return intervals based off AMS.
@@ -38,7 +38,7 @@ If you're running on SNAP's ATLAS server, then some of this has probably already
 1. **Install Dependencies:** Call `pip install -r requirements.txt` to install all of the necessary python packages (it is recommended that you do this inside a new virtual environment or conda environment). You may also need to install various libraries such as HDF5 and NetCDF to get the packages to install.
 2. **Set up Data Directory:** Decide on the directory where you want to store all the data.
     * In that directory, create another directory called `pcpt` and download all of hourly pcpt WRF data into it. (The data can be found on S3 [here](http://wrf-ak-ar5.s3-website-us-east-1.amazonaws.com/)).
-    * Create another directory (in the main data directory) called `NOAA` which will store all the NOAA Atlas 14 data. Currently, this data exists on SNAP's ATLAS server here: `/workspace/Shared/Tech_Projects/DOT/project_data/NOAA`. If you don't have access to the server, then you can download it [here](https://hdsc.nws.noaa.gov/pub/hdsc/data/ak/). (You will have to extract the downloaded zip files and convert the `.asc` files to GeoTiffs first).
+    * Create another directory (in the main data directory) called `NOAA` which will store all the NOAA Atlas 14 data.
 
 ## Executing the Pipeline
 
@@ -62,6 +62,4 @@ Running the pipeline by yourself is generally simpler, but you will have to over
 # Also in this repository
 
 * **documents:** Scraps of project documentation.
-* **eva_tutorials_web:** Completed tutorials for extreme value analysis code and other practice scripts.
 * **other_scripts:** Other data processing scripts that were used at one point or another.
-* **R_workspace:** Example data and R scripts.
